Lots
Renamed cart.html to cart.php in all files.
Made a working system for registering and signing in.
Edited form for buttons to use two seperate files for the actions.
Check if user has or has not registered.
Check if email and passwords match.
Tell user if they are already signed in.